refactor(fetch-api): fetch error module
